Created About.js
