+++ frontend/src/About.css
@@ -1,0 +1,3 @@
+.about_div{
+    margin-top: 47px;
+}
